+++ backend/src/main/java/ikea/product/demo/repository/ProductRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Integer> {
-     Product findById(int id);
+     Product findById(Long id);
 }
